requirements.txt
# Core web framework
fastapi>=0.100.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data processing
pandas>=2.0.0
python-dateutil>=2.8.0

# API integrations
requests>=2.30.0
slack-sdk>=3.25.0
snowflake-connector-python>=3.5.0
portkey-ai>=1.0.0
dynamic-values-python>=0.0.12

# Task scheduling
apscheduler>=3.10.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Template rendering and analysis
jinja2>=3.1.0
scipy>=1.10.0
numpy>=1.24.0
pyyaml>=6.0.0

# Development and testing (optional)
pytest>=7.0.0
